backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
tensorflow==2.14.0
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
